# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "健身计划管理系统API",
        "title": "Flow Link Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/plans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的健身计划列表，支持分页和状态筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "获取健身计划列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "计划状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.PaginatedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/plans/custom": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建完全自定义的健身计划",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "创建自定义健身计划",
                "parameters": [
                    {
                        "description": "自定义计划信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateCustomPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.FitnessPlan"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/plans/from-template": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "使用计划模板创建新的健身计划",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "基于模板创建健身计划",
                "parameters": [
                    {
                        "description": "模板创建请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreatePlanFromTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.FitnessPlan"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/plans/{planId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取健身计划的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "获取健身计划详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "计划ID",
                        "name": "planId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.FitnessPlan"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "计划ID不能为空",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "健身计划不存在",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定ID的健身计划",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "删除健身计划",
                "parameters": [
                    {
                        "type": "string",
                        "description": "计划ID",
                        "name": "planId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "计划ID不能为空",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/plans/{planId}/complete-day": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "标记健身计划中的某一天为已完成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "完成训练日",
                "parameters": [
                    {
                        "type": "string",
                        "description": "计划ID",
                        "name": "planId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "完成日期信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CompleteDayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "完成成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.FitnessPlan"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/plans/{planId}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新健身计划的状态（激活、暂停、完成等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健身计划"
                ],
                "summary": "更新计划状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "计划ID",
                        "name": "planId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdatePlanStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CompleteDayRequest": {
            "type": "object",
            "required": [
                "dayNumber"
            ],
            "properties": {
                "dayNumber": {
                    "type": "integer"
                },
                "recordId": {
                    "type": "integer"
                }
            }
        },
        "domain.CreateCustomPlanRequest": {
            "type": "object",
            "required": [
                "durationWeeks",
                "goal",
                "name",
                "startDate",
                "trainingDays",
                "trainingDaysPerWeek"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "durationWeeks": {
                    "type": "integer"
                },
                "goal": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "trainingDays": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TrainingDay"
                    }
                },
                "trainingDaysPerWeek": {
                    "type": "integer"
                }
            }
        },
        "domain.CreatePlanFromTemplateRequest": {
            "type": "object",
            "required": [
                "startDate",
                "templateId"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "templateId": {
                    "type": "integer"
                }
            }
        },
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.FitnessPlan": {
            "type": "object",
            "properties": {
                "completedDays": {
                    "description": "已完成的训练日",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "completionRate": {
                    "description": "完成率(百分比)",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentDay": {
                    "description": "当前第几天",
                    "type": "integer"
                },
                "currentWeek": {
                    "description": "当前第几周",
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "durationWeeks": {
                    "description": "计划周期(周)",
                    "type": "integer"
                },
                "endDate": {
                    "description": "结束日期 YYYY-MM-DD",
                    "type": "string"
                },
                "goal": {
                    "description": "训练目标",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startDate": {
                    "description": "开始日期 YYYY-MM-DD",
                    "type": "string"
                },
                "status": {
                    "description": "进行中/已完成/已暂停/已归档",
                    "type": "string"
                },
                "templateId": {
                    "description": "模板ID(0表示自定义)",
                    "type": "integer"
                },
                "totalCompletedDays": {
                    "description": "累计完成天数",
                    "type": "integer"
                },
                "trainingDays": {
                    "description": "训练日程",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TrainingDay"
                    }
                },
                "trainingDaysPerWeek": {
                    "description": "每周训练天数",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "domain.PaginatedData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "plans": {
                    "description": "用于计划"
                },
                "records": {
                    "description": "用于训练记录"
                },
                "templates": {
                    "description": "用于模板"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.TrainingDay": {
            "type": "object",
            "properties": {
                "dayName": {
                    "description": "训练日名称",
                    "type": "string"
                },
                "dayNumber": {
                    "description": "第几天",
                    "type": "integer"
                },
                "exercises": {
                    "description": "当日训练项目",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Exercise"
                    }
                },
                "isRestDay": {
                    "description": "是否为休息日",
                    "type": "boolean"
                },
                "notes": {
                    "description": "当日备注",
                    "type": "string"
                }
            }
        },
...
        "domain.UpdatePlanStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}